projects/whatsapp-automation-bot/requirements.txt
Flask==2.3.3
openai==1.3.0
psycopg2-binary==2.9.7
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.0
flask-cors==4.0.0
gunicorn==21.2.0
python-whatsapp-sdk==0.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.35.0
torch==2.1.0
